+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 We use [semantic versioning](http://semver.org/):
 
 - MAJOR version when you make incompatible API changes,
-- MINOR version when you add functionality in a backwards compatible manner, and
+- MINOR version when you add functionality in a backwards-compatible manner, and
 - PATCH version when you make backwards compatible bug fixes.
 
 # Next Release
@@ -12,6 +12,8 @@ We use [semantic versioning](http://semver.org/):
 - [fix] Remove retry logic for impacted tests request
 - [fix] Resolve possible memory leak during report generation
 - [feature] Use git.commit.id from git.properties instead of branch and timestamp
+- [breaking change] Reduce XML report size by only including source file coverage, no class coverage
+- [feature] New option `--ignore-uncovered-classes` to further reduce size of XML reports
 
 # 15.5.0
 - [feature] add TIA client library for integrating TIA in your custom test framework